pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grobid-metadata-enricher"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
]

[project.scripts]
grobid-metadata-enricher = "grobid_metadata_enricher.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pylint>=3.0",
    "pyarrow>=14.0",
    "pytest>=8.0",
    "httpx>=0.27",
]
bench = [
    "huggingface_hub>=0.24",
    "pyarrow>=15",
    "numpy>=1.26",
    "scipy>=1.12",
    "PyYAML>=6",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pylint.main]
py-version = "3.10"

[tool.pylint.format]
max-line-length = 120

[tool.pylint."messages control"]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-many-locals",
    "too-many-branches",
    "too-many-statements",
    "too-many-instance-attributes",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-few-public-methods",
    "broad-exception-caught",
    "import-outside-toplevel",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true
